+++ packages/expo-dev-launcher/android/src/debug/java/expo/modules/devlauncher/launcher/DevLauncherClientHost.kt
@@ -5,8 +5,10 @@ import com.facebook.react.ReactApplication
 import com.facebook.react.ReactNativeHost
 import com.facebook.react.ReactPackage
 import com.facebook.react.ReactPackageTurboModuleManagerDelegate
+import com.facebook.react.bridge.JSIModulePackage
 import com.facebook.react.bridge.JavaScriptExecutorFactory
 import com.facebook.react.config.ReactFeatureFlags
+import com.facebook.react.defaults.DefaultJSIModulePackage
 import com.facebook.react.shell.MainReactPackage
 import devmenu.com.th3rdwave.safeareacontext.SafeAreaContextPackage
 import expo.modules.adapters.react.ModuleRegistryAdapter
@@ -79,4 +81,11 @@ class DevLauncherClientHost(
     method.isAccessible = true
     return method.invoke(appHost) as ReactPackageTurboModuleManagerDelegate.Builder
   }
+
+  override fun getJSIModulePackage(): JSIModulePackage? =
+    if (ReactFeatureFlags.enableFabricRenderer) {
+      DefaultJSIModulePackage(this)
+    } else {
+      null
+    }
 }